Python_Backend/poc_snow_flake/requirements.txt
Flask-Cors
flask
mysql
pytz
pandas
mysql-connector-python
python-dateutil
requests
snowflake-connector-python
python-dotenv
flask_mail